Update Share functionality in unique token expanded state
Since some networks are not supported by Rainbow web yet,
we want to fallback to using the permalink as the Share content.
If a permalink also does not exist, then we do not show the Share
button.
